requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
faiss-cpu>=1.7.4
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
numpy>=1.24.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
pytest>=7.4.0
